requirements.txt
python-telegram-bot
openai-whisper
google-generativeai
python-dotenv
httpx
pytest
pytest-asyncio
gitpython